+++ t/Rgba.t
@@ -15,25 +15,30 @@ my Gnome::Gdk3::RGBA $r;
 subtest 'ISA test', {
   $r .= new( :red(.5e0), :green('0.5'), :blue(0.5), :alpha(.5e0));
   isa-ok $r, Gnome::Gdk3::RGBA, '.new(:red,...)';
+  ok $r.is-valid, 'r,g,b,a is valid';
 
-  my Gnome::Gdk3::RGBA $c1 .= new(:rgba($r()));
-  isa-ok $c1, Gnome::Gdk3::RGBA, '.new(:rgba(N-GdkRGBA))';
+  my Gnome::Gdk3::RGBA $c1 .= new(:native-object($r()));
+  isa-ok $c1, Gnome::Gdk3::RGBA, '.new(:native-object(N-GdkRGBA))';
   is $c1.to-string, 'rgba(128,128,128,0.5)', 'compare string ok';
+  ok $r.is-valid, 'rgba is valid';
 
-  $c1 .= new(:rgba($r));
-  isa-ok $c1, Gnome::Gdk3::RGBA, '.new(:rgba(Gnome::Gdk3::RGBA))';
+  $c1 .= new(:native-object($r));
+  isa-ok $c1, Gnome::Gdk3::RGBA, '.new(:native-object(Gnome::Gdk3::RGBA))';
   is $c1.to-string, 'rgba(128,128,128,0.5)', 'compare string ok';
 
   $c1 = Gnome::Gdk3::RGBA.new(:rgba<#ff0>);
   isa-ok $c1, Gnome::Gdk3::RGBA, '.new(:rgba(Str))';
   is $c1.to-string, 'rgb(255,255,0)', 'compare string ok';
+
+  $c1.clear-object;
+  nok $c1.is-valid, '.clear-object()';
 }
 
 #-------------------------------------------------------------------------------
 subtest 'Manipulations', {
   is $r.to-string, 'rgba(128,128,128,0.5)', '.to-string()';
 
-  my Gnome::Gdk3::RGBA $r2 .= new(:rgba($r.gdk-rgba-copy));
+  my Gnome::Gdk3::RGBA $r2 .= new(:native-object($r.gdk-rgba-copy));
   is $r2.to-string, 'rgba(128,128,128,0.5)', '.gdk-rgba-copy()';
 
 
@@ -41,7 +46,7 @@ subtest 'Manipulations', {
   my UInt $key = $r.gdk-rgba-hash;
 #note "k: $key";
   ok $key > 0, '.gdk-rgba-hash(Gnome::Gdk3::RGBA)';
-  my Gnome::Gdk3::RGBA $r3 .= new(:rgba($r2.gdk-rgba-hash($key)));
+  my Gnome::Gdk3::RGBA $r3 .= new(:native-object($r2.gdk-rgba-hash($key)));
   is $r.to-string, $r3.to-string, '.gdk-rgba-hash(UInt)';
 
   ok $r.gdk-rgba-equal($r3), '.gdk-rgba-equal(): =';